added a utils3.c file
